# docs/swagger/agent.yml
openapi: 3.0.0
info:
  title: RSS爬虫代理API
  description: RSS爬虫代理相关的API接口
  version: 1.0.0
tags:
  - name: 爬虫代理
    description: RSS爬虫代理相关操作
paths:
  /register:
    post:
      summary: 注册爬虫代理
      description: 注册新的爬虫代理
      tags:
        - 爬虫代理
      security:
        - appKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                agent_id:
                  type: string
                  description: 代理唯一标识，可选，不提供则使用hostname
                  example: "crawler-agent-1"
                hostname:
                  type: string
                  description: 主机名
                  example: "crawler-server-01"
                ip_address:
                  type: string
                  description: IP地址
                  example: "192.168.1.100"
                version:
                  type: string
                  description: 爬虫版本
                  example: "1.0.0"
                capabilities:
                  type: object
                  description: 爬虫能力
                  example: {"js_rendering": true, "proxy_support": true}
      responses:
        200:
          description: 代理注册成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "代理注册成功"
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        example: "a1b2c3d4e5f6"
                      agent_id:
                        type: string
                        example: "crawler-agent-1"
                      hostname:
                        type: string
                        example: "crawler-server-01"
                      ip_address:
                        type: string
                        example: "192.168.1.100"
                      version:
                        type: string
                        example: "1.0.0"
                      capabilities:
                        type: object
                        example: {"js_rendering": true, "proxy_support": true}
                      status:
                        type: integer
                        example: 1
                      last_heartbeat:
                        type: string
                        format: date-time
                        example: "2023-01-01T12:00:00"
                      created_at:
                        type: string
                        format: date-time
                        example: "2023-01-01T12:00:00"
  /heartbeat:
    post:
      summary: 更新代理心跳
      description: 更新代理心跳状态和性能信息
      tags:
        - 爬虫代理
      security:
        - appKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - agent_id
              properties:
                agent_id:
                  type: string
                  description: 代理唯一标识
                  example: "crawler-agent-1"
                status:
                  type: integer
                  description: 状态：1=活跃, 2=闲置, 3=离线
                  example: 1
                performance:
                  type: object
                  description: 性能指标
                  example: {
                    "total_tasks": 100,
                    "success_tasks": 90,
                    "failed_tasks": 10,
                    "avg_processing_time": 2.5
                  }
      responses:
        200:
          description: 心跳更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "心跳更新成功"
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        example: "a1b2c3d4e5f6"
                      agent_id:
                        type: string
                        example: "crawler-agent-1"
                      status:
                        type: integer
                        example: 1
                      last_heartbeat:
                        type: string
                        format: date-time
                        example: "2023-01-01T12:15:00"
                      performance:
                        type: object
                        example: {
                          "total_tasks": 100,
                          "success_tasks": 90,
                          "failed_tasks": 10,
                          "avg_processing_time": 2.5
                        }
  /list:
    get:
      summary: 获取代理列表
      description: 获取所有爬虫代理列表
      tags:
        - 爬虫代理
      security:
        - jwtAuth: []
      parameters:
        - name: status
          in: query
          description: 状态筛选：1=活跃, 2=闲置, 3=离线
          required: false
          schema:
            type: integer
            enum: [1, 2, 3]
      responses:
        200:
          description: 获取代理列表成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "成功"
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: "a1b2c3d4e5f6"
                        agent_id:
                          type: string
                          example: "crawler-agent-1"
                        hostname:
                          type: string
                          example: "crawler-server-01"
                        ip_address:
                          type: string
                          example: "192.168.1.100"
                        version:
                          type: string
                          example: "1.0.0"
                        capabilities:
                          type: object
                          example: {"js_rendering": true, "proxy_support": true}
                        status:
                          type: integer
                          example: 1
                        last_heartbeat:
                          type: string
                          format: date-time
                          example: "2023-01-01T12:15:00"
                        performance:
                          type: object
                          example: {
                            "total_tasks": 100,
                            "success_tasks": 90,
                            "failed_tasks": 10,
                            "avg_processing_time": 2.5
                          }
                        created_at:
                          type: string
                          format: date-time
                          example: "2023-01-01T12:00:00"
  /detail:
    get:
      summary: 获取代理详情
      description: 获取特定爬虫代理的详细信息
      tags:
        - 爬虫代理
      security:
        - jwtAuth: []
      parameters:
        - name: agent_id
          in: query
          description: 代理ID
          required: true
          schema:
            type: string
      responses:
        200:
          description: 获取代理详情成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "成功"
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        example: "a1b2c3d4e5f6"
                      agent_id:
                        type: string
                        example: "crawler-agent-1"
                      hostname:
                        type: string
                        example: "crawler-server-01"
                      ip_address:
                        type: string
                        example: "192.168.1.100"
                      version:
                        type: string
                        example: "1.0.0"
                      capabilities:
                        type: object
                        example: {"js_rendering": true, "proxy_support": true}
                      status:
                        type: integer
                        example: 1
                      last_heartbeat:
                        type: string
                        format: date-time
                        example: "2023-01-01T12:15:00"
                      performance:
                        type: object
                        example: {
                          "total_tasks": 100,
                          "success_tasks": 90,
                          "failed_tasks": 10,
                          "avg_processing_time": 2.5
                        }
                      crawl_stats:
                        type: object
                        example: {
                          "total_articles": 500,
                          "success_rate": 95.2,
                          "avg_crawl_time": 2.3
                        }
                      created_at:
                        type: string
                        format: date-time
                        example: "2023-01-01T12:00:00"
                      updated_at:
                        type: string
                        format: date-time
                        example: "2023-01-01T12:15:00"